This commit contains the following improvements:
a) Created the documentmodel package that now encapsulates the MUCKE
document model consisting of a list of facets that each can be
transformed into a list of concepts.

b) Created an initial form of the DocumentIndexer implementation (with
additional improvements pending). The DocumentIndexer creates a
meta-index on document level (as in the MUCKE document model defined in
the documentmodel package). The DocumentIndexer is triggered by the
IndexManager. The DocumentIndexer is fully configurable by the
configuration file. The DocumentIndexer stores a list of facet ids in a
database table, which is stored in the system database (for now, at
least) (see data package).

c) The index package now has a set of FacetIndexer implementations, for
text, tag and image (currently only implemented for the text facet, see
StandardTextFacetIndexer).

d) The ConfigurationManager provides access to the DBManager object. 

e) Bugfix: The DBManager provides a single connection to the database
and does not create a new connection every time a query is processed.
This caused unexpected crashes after 5 minutes of operation earlier

f) Adapted documentation of the system architecture and the index
package in particular.

g) Added library of Apache IO commons: commons-io-2.4.jar that handles
all kinds of input/output operations and is used by the indexing package
to further standardize code.

TODO Basics:

1) Implement StandardTagFacetIndexer
2) Implement StandardImageFacetIndexer with a most basic form of Lire

TODO Optimizations:

1) Improve DBManager to use prepared statements, current indexing of
DocumentIndexer takes ~30 minutes. Should become a lot faster.

2) Implement Saxon 9 in XMLTools to allow for XPath 2.0 expressions in
configuration. This would allow more powerful expressions for extracting
text fractions from XML files throughout the system architecture.