Main modification: Now, we have an added Lire-based ImageFacetIndexer that indexes directories of images recursively.
Detailed modfications:

1) ConfigConstants have been extended with parameters valid for the
StandardImageIndexer (e.g. the feature list of color measure and color
distribution features that can be used for image indexing).
2) ConfigurationManager now creates, more precicely than before,
FacetIndexers based on an updated constructor signature.
3) Both TextFacetIndexer and ImageFacetIndexer now store their index as
defined in configuration file.

TODO for the index package:

a) Implement StandardTagFacetIndexer. What exactly is different between
Tag and Text w.r.t the FacetIndexer? Just the less transforming Analyzer
(e.g. by not stemming words, but instead using them as categories
without modification?) 

b) Change StandardImageFacetIndexer to work recursively on directories
instead of read-in filename as Strings. That avoids the long waiting
time before it starts and the need to modify logic ot fit this. When
using files, it should be possible to encapsulate the facetId extraction
in a method and reuse it for all three StandardFacetIndexer
implementations.